examples/rest/requirements.txt
Django==3.2.14
djangorestframework==3.12.4
passlib==1.7.4
